feat(system): remove the old JDBC queue
