--- jdbc-h2/src/main/java/io/kestra/runner/h2/H2Queue.java
@@ -1,79 +1,0 @@
-package io.kestra.runner.h2;
-
-import io.kestra.jdbc.repository.AbstractJdbcRepository;
-import io.kestra.jdbc.runner.JdbcQueue;
-import io.micronaut.context.ApplicationContext;
-import org.jooq.*;
-import org.jooq.Record;
-import org.jooq.impl.DSL;
-import org.jooq.impl.SQLDataType;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-public class H2Queue<T> extends JdbcQueue<T> {
-    public H2Queue(Class<T> cls, ApplicationContext applicationContext) {
-        super(cls, applicationContext);
-    }
-
-    @Override
-    protected Condition buildTypeCondition(String type) {
-        return AbstractJdbcRepository.field("type").eq(type);
-    }
-
-    @Override
-    protected Result<Record> receiveFetch(DSLContext ctx, String consumerGroup, String queueType, boolean forUpdate) {
-        var select =  ctx.select(
-                AbstractJdbcRepository.field("value"),
-                AbstractJdbcRepository.field("offset")
-            )
-            .from(this.table)
-            .where(AbstractJdbcRepository.field("type").eq(queueType()))
-            .and(DSL.or(List.of(
-                AbstractJdbcRepository.field("consumers").isNull(),
-                DSL.condition("NOT(ARRAY_CONTAINS(\"consumers\", ?))", queueType)
-            )));
-
-        if (consumerGroup != null) {
-            select = select.and(AbstractJdbcRepository.field("consumer_group").eq(consumerGroup));
-        } else {
-            select = select.and(AbstractJdbcRepository.field("consumer_group").isNull());
-        }
-
-        var limitSelect = select
-            .orderBy(AbstractJdbcRepository.field("offset").asc())
-            .limit(configuration.pollSize());
-        ResultQuery<Record2<Object, Object>> configuredSelect = limitSelect;
-
-        if (forUpdate) {
-            configuredSelect = limitSelect.forUpdate().skipLocked();
-        }
-
-        return configuredSelect
-            .fetchMany()
-            .getFirst();
-    }
-
-    @Override
-    protected void doUpdateGroupOffsets(DSLContext ctx, String consumerGroup, String queueType, List<Integer> offsets) {
-        var update = ctx.update(DSL.table(table.getName()))
-            .set(
-                AbstractJdbcRepository.field("consumers"),
-                DSL.field(
-                    "ARRAY_APPEND(COALESCE(\"consumers\", ARRAY[]), ?)",
-                    SQLDataType.VARCHAR(50).getArrayType(),
-                    (Object) new String[]{queueType}
-                )
-            )
-            .set(AbstractJdbcRepository.field("updated"), LocalDateTime.now())
-            .where(AbstractJdbcRepository.field("offset").in(offsets.toArray(Integer[]::new)));
-
-        if (consumerGroup != null) {
-            update = update.and(AbstractJdbcRepository.field("consumer_group").eq(consumerGroup));
-        } else {
-            update = update.and(AbstractJdbcRepository.field("consumer_group").isNull());
-        }
-
-        update.execute();
-    }
-}
