--- jdbc-h2/src/main/java/io/kestra/runner/h2/H2QueueEnabled.java
@@ -1,14 +1,0 @@
-package io.kestra.runner.h2;
-
-import io.micronaut.context.annotation.DefaultImplementation;
-import io.micronaut.context.annotation.Requires;
-
-import java.lang.annotation.*;
-
-@Documented
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.PACKAGE, ElementType.TYPE})
-@Requires(property = "kestra.queue.type", pattern = "h2|memory")
-@DefaultImplementation
-public @interface H2QueueEnabled {
-}
