--- jdbc-h2/src/test/java/io/kestra/runner/h2/H2JdbcCleanerTest.java
@@ -1,6 +1,0 @@
-package io.kestra.runner.h2;
-
-import io.kestra.jdbc.runner.AbstractJdbcCleanerTest;
-
-public class H2JdbcCleanerTest extends AbstractJdbcCleanerTest {
-}
