--- jdbc-mysql/src/main/java/io/kestra/runner/mysql/MysqlQueueEnabled.java
@@ -1,14 +1,0 @@
-package io.kestra.runner.mysql;
-
-import io.micronaut.context.annotation.DefaultImplementation;
-import io.micronaut.context.annotation.Requires;
-
-import java.lang.annotation.*;
-
-@Documented
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.PACKAGE, ElementType.TYPE})
-@Requires(property = "kestra.queue.type", value = "mysql")
-@DefaultImplementation
-public @interface MysqlQueueEnabled {
-}
