--- jdbc-mysql/src/test/java/io/kestra/runner/mysql/MysqlJdbcCleanerTest.java
@@ -1,6 +1,0 @@
-package io.kestra.runner.mysql;
-
-import io.kestra.jdbc.runner.AbstractJdbcCleanerTest;
-
-public class MysqlJdbcCleanerTest extends AbstractJdbcCleanerTest {
-}
