--- jdbc-postgres/src/main/java/io/kestra/runner/postgres/PostgresJdbcCleanerService.java
@@ -1,15 +1,0 @@
-package io.kestra.runner.postgres;
-
-import io.kestra.jdbc.runner.JdbcCleanerService;
-import jakarta.inject.Singleton;
-import org.jooq.Condition;
-import org.jooq.impl.DSL;
-
-@Singleton
-@PostgresQueueEnabled
-public class PostgresJdbcCleanerService implements JdbcCleanerService {
-    @Override
-    public Condition buildTypeCondition(String type) {
-        return DSL.condition("type = CAST(? AS queue_type)", type);
-    }
-}
