--- jdbc-postgres/src/main/java/io/kestra/runner/postgres/PostgresQueueEnabled.java
@@ -1,14 +1,0 @@
-package io.kestra.runner.postgres;
-
-import io.micronaut.context.annotation.DefaultImplementation;
-import io.micronaut.context.annotation.Requires;
-
-import java.lang.annotation.*;
-
-@Documented
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.PACKAGE, ElementType.TYPE})
-@Requires(property = "kestra.queue.type", value = "postgres")
-@DefaultImplementation
-public @interface PostgresQueueEnabled {
-}
