--- jdbc-postgres/src/test/java/io/kestra/runner/postgres/PostgresJdbcCleanerTest.java
@@ -1,6 +1,0 @@
-package io.kestra.runner.postgres;
-
-import io.kestra.jdbc.runner.AbstractJdbcCleanerTest;
-
-public class PostgresJdbcCleanerTest extends AbstractJdbcCleanerTest {
-}
