+++ jdbc/src/main/java/io/kestra/jdbc/runner/JdbcQueueConfiguration.java
@@ -4,9 +4,6 @@ import io.micronaut.context.annotation.ConfigurationProperties;
 import io.micronaut.core.bind.annotation.Bindable;
 
 import java.time.Duration;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
 
 @ConfigurationProperties("kestra.jdbc.queues")
 public record JdbcQueueConfiguration(
@@ -23,36 +20,4 @@ public record JdbcQueueConfiguration(
     @Bindable(defaultValue = "true")
     Boolean immediateRepoll
 ) {
-
-    public List<Step> computeSteps() {
-        if (this.maxPollInterval.compareTo(this.minPollInterval) < 0) {
-            throw new IllegalArgumentException("'maxPollInterval' (" + this.maxPollInterval + ") must be greater than or equal to 'minPollInterval' (" + this.minPollInterval + ")");
-        }
-
-        if (this.maxPollInterval.equals(this.minPollInterval)) {
-            return List.of(new Step(this.minPollInterval, Duration.ZERO));
-        }
-
-        List<Step> steps = new ArrayList<>();
-        Step currentStep = new Step(this.maxPollInterval, this.pollSwitchInterval);
-        steps.add(currentStep);
-        for (int i = 0; i < switchSteps; i++) {
-            Duration stepPollInterval = Duration.ofMillis(currentStep.pollInterval().toMillis() / 2);
-            if (stepPollInterval.compareTo(minPollInterval) < 0) {
-                stepPollInterval = minPollInterval;
-            }
-            Duration stepSwitchInterval = Duration.ofMillis(currentStep.switchInterval().toMillis() / 2);
-            currentStep = new Step(stepPollInterval, stepSwitchInterval);
-            steps.add(currentStep);
-        }
-        Collections.sort(steps);
-        return steps;
-    }
-
-    public record Step(Duration pollInterval, Duration switchInterval) implements Comparable<Step> {
-        @Override
-        public int compareTo(Step o) {
-            return this.switchInterval.compareTo(o.switchInterval);
-        }
-    }
 }
